+++ src/app/movie-card/movie-card.component.scss
@@ -1,0 +1,4 @@
+.movieCardImage {
+    max-height: 500px; 
+    max-width: 500px;
+}
