+++ src/app/movie-card/movie-card.component.ts
@@ -29,6 +29,11 @@ getMovies(): void {
       console.log(this.movies);
       return this.movies;
     });
+  };
+
+  isFav(movie: any): boolean {
+
+    return this.favoriteMoviesIDs.includes(movie._id);
   }
 
 // This is the function that will open the dialog when the director button is clicked  
@@ -54,4 +59,27 @@ openSynopsisDialog(synopsis): void {
     width: '280px'
   });
 };
+
+// This is the function that toggles the movie in the user's favorites  
+toggleFav(movie: any): void {
+  console.log("toggleFav", movie)
+  const isFavorite = this.isFav(movie);
+    this.addFavMovies(movie);
+};
+
+// This is the function that will add the movie to the user's favorites when the heart button is clicked  
+addFavMovies(movie: any): void {
+  console.log("addFavMovies")
+  const user = JSON.parse(localStorage.getItem('user') || '{}');
+  if (user) {
+    console.log("if statement")
+    this.movieAPIData.addFavoriteMovies(user.Username, movie._id).subscribe((result) => {
+      localStorage.setItem('user', JSON.stringify(result));
+      this.getFavMovies(); // Refresh favorite movies after adding a new one
+      this.snackBar.open(`${movie.movieName} has been added to your favorites`, 'OK', {
+        duration: 1000,
+      });
+    });
+  }
+}
 }
