+++ src/app/user-profile/user-profile.component.html
@@ -45,12 +45,12 @@
 
 
                     <!-- Update and Delete User buttons -->
-                    <!-- <div class="button-container">
+                    <div class="button-container">
                         <button mat-raised-button type="submit" color="primary">Update</button>
-                        <button mat-raised-button color="warn" (click)="deleteUser()">
-                            Permanently delete your account
-                        </button>
-                    </div> -->
+                        <!-- <button mat-raised-button color="warn" (click)="deleteUser()"> -->
+                        <!-- Permanently delete your account
+                        </button> -->
+                    </div>
                 </form>
             </div>
         </div>
